+++ packages/explorer-ui/next.config.js
@@ -2,6 +2,9 @@
 const nextConfig = {
   reactStrictMode: true,
   webpack: (config, { isServer }) => {
+    if (!isServer) {
+      config.output.publicPath = '/assets'
+    }
     config.resolve.extensions = ['.js', '.jsx', '.ts', '.tsx', '.json']
     config.stats = {
       warnings: true,
@@ -9,6 +12,10 @@ const nextConfig = {
       errorDetails: true,
       modules: true,
     }
+    config.module.rules.push({
+      test: /\.tsx?$/,
+      use: 'ts-loader',
+    })
     return config
   },
   eslint: {